Use long, not int.
git-svn-id: https://svn.apache.org/repos/asf/jena/trunk@1393488 13f79535-47bb-0310-9956-ffa450edef68

+++ jena-arq/src/main/java/org/openjena/riot/web/HttpOp.java
@@ -132,7 +132,7 @@ public class HttpOp
      */
    //TODO Use MediaType
     public static void execHttpPost(String url, String contentType, 
-                                    InputStream input, int length)
+                                    InputStream input, long length)
     {
         execHttpPost(url, contentType, input, length, null, null ) ;
     }
@@ -158,7 +158,7 @@ public class HttpOp
      * The input stream is assumed to be UTF-8.
      */
     public static void execHttpPost(String url, String contentType, 
-                                    InputStream input, int length,
+                                    InputStream input, long length,
                                     String acceptType, Map<String, HttpResponseHandler> handlers)
     {
         
@@ -246,7 +246,7 @@ public class HttpOp
     }
     
     /** Execute an HTTP PUT operation */
-    public static void execHttpPut(String url, String contentType, InputStream input, int length)
+    public static void execHttpPut(String url, String contentType, InputStream input, long length)
     {
         InputStreamEntity e = new InputStreamEntity(input, length) ;
         e.setContentType(contentType) ;
